requirements.txt
# Core data
yfinance>=0.2.38
pandas>=2.1.0
numpy>=1.26.0
requests>=2.31.0

# Technical indicators
pandas-ta>=0.3.14b

# Database
sqlalchemy>=2.0.0

# Visualization
matplotlib>=3.8.0
plotly>=5.18.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.29.0

# Phase 2 — NLP & Sentiment
transformers>=4.38.0
torch>=2.2.0
newsapi-python>=0.2.7
python-dotenv>=1.0.0

# Phase 3 — ML models
xgboost>=2.0.0
scikit-learn>=1.4.0
joblib>=1.3.0
imbalanced-learn>=0.12.0

# Phase 4 — API backend
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9